requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.10.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==5.1.3
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
Flask==3.0.3
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
plotly==5.24.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rpds-py==0.21.0
six==1.16.0
sqlparse==0.5.1
swagger-spec-validator==3.0.4
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
